fastapi_authentications/jwt/requirements.txt
fastapi 
uvicorn 
python-jose[cryptography]
passlib[bcrypt]